chore: Change package structure
