+++ src/test/kotlin/lotto/LottoNumberGeneratorTest.kt
@@ -3,11 +3,12 @@ package lotto
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.ints.shouldBeInRange
 import io.kotest.matchers.shouldBe
+import lotto.util.LottoNumberGenerator
 
 class LottoNumberGeneratorTest : StringSpec({
     "1~45 사이의 랜덤 숫자 6개를 반환한다."{
         //when
-        val lottoNumbers = LottoNumberGenerator.generateNumbers()
+        val lottoNumbers = LottoNumberGenerator.generateNumbers().toList()
         //then
         lottoNumbers.size shouldBe 6
         lottoNumbers[0] shouldBeInRange 1..45
