+++ src/test/kotlin/lotto/WinnerTicketTest.kt
@@ -2,6 +2,8 @@ package lotto
 
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
+import lotto.domain.LottoTicket
+import lotto.domain.WinnerTicket
 
 class WinnerTicketTest : BehaviorSpec({
